src/backend/requirements.txt
anyio==4.8.0
bcrypt==4.2.0
blinker==1.9.0
cachelib==0.13.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.7
cryptography==43.0.3
Flask==3.1.0
Flask-MySQLdb==2.0.0
Flask-Session==0.8.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
msgspec==0.18.6
mysql-connector-python==9.1.0
mysqlclient==2.2.6
pycparser==2.22
PyMySQL==1.1.1
python-decouple==3.8
python-telegram-bot==21.10
requests==2.32.3
sniffio==1.3.1
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
